+++ frontend/src/locales/lang/en-US.js
@@ -52,7 +52,12 @@ const locale = {
   'app.setting.copy': 'Copy Setting',
   'app.setting.loading': 'Loading theme',
   'app.setting.copyinfo': 'copy success，please replace defaultSettings in src/models/setting.js',
-  'app.setting.production.hint': 'Setting panel shows in development environment only, please manually modify'
+  'app.setting.production.hint': 'Setting panel shows in development environment only, please manually modify',
+
+  'menu.setting.resolve': 'Resolve',
+  'Host': 'Host',
+  'Type': 'Type',
+  'Value': 'Value'
 }
 
 export default {